Move 'here()' cli impl to screen code
