+++ javarosa/util/schema-gen/src/org/javarosa/engine/FunctionExtensions.java
@@ -105,35 +105,4 @@ public class FunctionExtensions {
         }
     }
 
-    public static class HereDummyFunc implements IFunctionHandler {
-        private final double lat;
-        private final double lon;
-
-        public HereDummyFunc(double lat, double lon) {
-            this.lat = lat;
-            this.lon = lon;
-        }
-
-        @Override
-        public String getName() {
-            return "here";
-        }
-
-        @Override
-        public Vector getPrototypes() {
-            Vector<Class[]> p = new Vector<>();
-            p.addElement(new Class[0]);
-            return p;
-        }
-
-        @Override
-        public boolean rawArgs() {
-            return false;
-        }
-
-        @Override
-        public Object eval(Object[] args, EvaluationContext ec) {
-            return new GeoPointData(new double[]{lat, lon, 0, 10}).getDisplayText();
-        }
-    }
 }
